# cpf-biz-admin/frontend/openapi/cpf-openapi.json
{
  "components": {
    "schemas": {
...
      "CpfControllerSourceResponse": {
        "additionalProperties": true,
        "type": "object"
      },
      "CpfSensitiveDataAccessRequest": {
        "additionalProperties": true,
        "description": "Controller-source pre-runtime placeholder; runtime OpenAPI must supply validation details.",
        "type": "object"
      },
...
      "EmployeeRequest": {
        "additionalProperties": true,
        "description": "Controller-source pre-runtime placeholder; runtime OpenAPI must supply validation details.",
        "type": "object"
      },
...
      "OrganizationRequest": {
        "additionalProperties": true,
        "description": "Controller-source pre-runtime placeholder; runtime OpenAPI must supply validation details.",
        "type": "object"
      },
...
      "CpfApiError": {
        "type": "object",
        "required": [
          "status",
          "code",
          "message",
          "path"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "transactionId": {
            "type": "string",
            "minLength": 34,
            "maxLength": 34
          }
        }
      }
    },
    "responses": {
      "CpfUnauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CpfApiError"
            }
          }
        }
      },
      "CpfForbidden": {
        "description": "Permission denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CpfApiError"
            }
          }
        }
      },
      "CpfNotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CpfApiError"
            }
          }
        }
      },
      "CpfConflict": {
        "description": "State/version conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CpfApiError"
            }
          }
        }
      },
      "CpfUnprocessable": {
        "description": "Validation or semantic request failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CpfApiError"
            }
          }
        }
      },
      "CpfTooManyRequests": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CpfApiError"
            }
          }
        }
      },
      "CpfInternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CpfApiError"
            }
          }
        }
      },
      "CpfUnavailable": {
        "description": "Service temporarily unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CpfApiError"
            }
          }
        }
      }
    }
  },
  "info": {
    "title": "CPF BZA controller source pre-runtime contract",
    "version": "0.0.0-pre-runtime"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/bza/backoffice/audits": {
      "get": {
        "operationId": "bzaBackofficeFindBusinessAudits",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      }
    },
    "/api/bza/backoffice/employees": {
      "get": {
        "operationId": "bzaBackofficeFindEmployees",
        "parameters": [
          {
            "in": "query",
            "name": "organizationCode",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      },
      "post": {
        "operationId": "bzaBackofficeSaveEmployee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          },
          "409": {
            "$ref": "#/components/responses/CpfConflict"
          },
          "422": {
            "$ref": "#/components/responses/CpfUnprocessable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      }
    },
    "/api/bza/backoffice/employees/page": {
      "get": {
        "operationId": "bzaBackofficeFindEmployeesPage",
        "parameters": [
          {
            "in": "query",
            "name": "organizationCode",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      }
    },
    "/api/bza/backoffice/employees/{employeeNo}/contacts/raw": {
      "post": {
        "operationId": "bzaBackofficeEmployeeRawContact",
        "parameters": [
          {
            "in": "path",
            "name": "employeeNo",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CpfSensitiveDataAccessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          },
          "404": {
            "$ref": "#/components/responses/CpfNotFound"
          },
          "409": {
            "$ref": "#/components/responses/CpfConflict"
          },
          "422": {
            "$ref": "#/components/responses/CpfUnprocessable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      }
    },
    "/api/bza/backoffice/organizations": {
      "get": {
        "operationId": "bzaBackofficeFindOrganizations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      },
      "post": {
        "operationId": "bzaBackofficeSaveOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          },
          "409": {
            "$ref": "#/components/responses/CpfConflict"
          },
          "422": {
            "$ref": "#/components/responses/CpfUnprocessable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      }
    },
    "/api/bza/backoffice/organizations/page": {
      "get": {
        "operationId": "bzaBackofficeFindOrganizationsPage",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      }
    },
    "/api/bza/backoffice/permissions/effective": {
      "get": {
        "operationId": "bzaBackofficeFindEffectivePermissions",
        "parameters": [
          {
            "in": "query",
            "name": "loginId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpfControllerSourceResponse"
                }
              }
            },
            "description": "Controller source contract response"
          },
          "401": {
            "$ref": "#/components/responses/CpfUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/CpfForbidden"
          },
          "429": {
            "$ref": "#/components/responses/CpfTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/CpfInternalError"
          },
          "503": {
            "$ref": "#/components/responses/CpfUnavailable"
          }
        },
        "x-cpf-controller-source": "cpf-biz-admin/src/main/java/com/cpf/bizadmin/backoffice/controller/BzaBackofficeController.java"
      }
    },
...
  },
...
}